Extract `@wp/api` module
This is just the `api` module for now, since all of that is WordPress-related
