+++ src/@wp/api/index.js
